combine yaw reg tests into one test file
